+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" >
+    xmlns:tools="http://schemas.android.com/tools">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -14,26 +14,21 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar"
-        tools:targetApi="31" >
+        tools:targetApi="31">
         <activity
-            android:name=".player.presentation.activity.MusicPlayerActivity"
-            android:exported="false" />
-        <activity
-            android:name=".media.presentation.activity.MediaLibraryActivity"
-            android:exported="false" />
-        <activity
-            android:name=".search.presentation.activity.SearchActivity"
-            android:exported="false" />
-        <activity
-            android:name=".settings.presentation.activity.SettingActivity"
-            android:exported="false" />
-        <activity
-            android:name=".main.presentation.activity.MainActivity"
-            android:exported="true" >
+            android:name=".ui.root.RootActivity"
+            android:exported="true"
+            android:windowSoftInputMode="adjustNothing">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".player.presentation.fragment.MusicPlayerFragment"
+            android:exported="false" />
+
     </application>
+
 </manifest>
