+++ app/src/main/java/com/example/playlistmaker/app/Extensions.kt
@@ -1,17 +1,16 @@
 package com.example.playlistmaker.app
 
 import android.content.Context
-import android.content.Intent
 import android.os.Build
 import android.os.Build.VERSION.SDK_INT
+import android.os.Bundle
 import android.os.Parcelable
 import android.util.TypedValue
 
 const val RADIUS_CUT_IMAGE = 10f
-
-inline fun <reified T : Parcelable> Intent.parcelable(key: String): T? = when {
-    SDK_INT >= Build.VERSION_CODES.TIRAMISU -> getParcelableExtra(key, T::class.java)
-    else -> @Suppress("DEPRECATION") getParcelableExtra(key) as? T
+inline fun <reified T : Parcelable> Bundle.parcelable(key: String): T? = when {
+    SDK_INT >= Build.VERSION_CODES.TIRAMISU -> getParcelable(key, T::class.java)
+    else -> @Suppress("DEPRECATION") getParcelable(key) as? T
 }
 
 fun dpToPx(dp: Float, context: Context): Int {
