+++ app/src/main/java/com/example/playlistmaker/di/ViewModelModule.kt
@@ -1,26 +1,20 @@
 package com.example.playlistmaker.di
 
-import com.example.playlistmaker.main.presentation.view_model.MainViewModel
+import com.example.playlistmaker.media.presentation.view_model.MediaLibraryViewModel
 import com.example.playlistmaker.player.presentation.view_model.MusicPlayerViewModel
-import com.example.playlistmaker.search.presentation.view_model.SearchActivityViewModel
+import com.example.playlistmaker.search.presentation.view_model.SearchFragmentViewModel
 import com.example.playlistmaker.settings.presentation.view_model.SettingsViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val viewModelModule = module {
 
-    viewModel<MainViewModel> {
-        MainViewModel(
-            settingsInteractor = get()
-        )
-    }
-
     viewModel<MusicPlayerViewModel> {
         MusicPlayerViewModel()
     }
 
-    viewModel<SearchActivityViewModel> {
-        SearchActivityViewModel(
+    viewModel<SearchFragmentViewModel> {
+        SearchFragmentViewModel(
             sharedPrefsInteractor = get(),
             tracksInteractor = get()
         )
@@ -31,5 +25,9 @@ val viewModelModule = module {
             sharingInteractor = get(),
             settingsInteractor = get()
         )
+    }
+
+    viewModel<MediaLibraryViewModel>{
+        MediaLibraryViewModel(get())
     }
 }
