+++ app/src/main/java/com/example/playlistmaker/media/presentation/fragment/MediaLibraryFragment.kt
@@ -1,0 +1,58 @@
+package com.example.playlistmaker.media.presentation.fragment
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.appcompat.app.AppCompatDelegate
+import androidx.fragment.app.Fragment
+import com.example.playlistmaker.R
+import com.example.playlistmaker.databinding.FragmentMediaLibraryBinding
+import com.example.playlistmaker.media.presentation.view_model.MediaLibraryViewModel
+import com.google.android.material.tabs.TabLayoutMediator
+import org.koin.androidx.viewmodel.ext.android.viewModel
+
+class MediaLibraryFragment:Fragment() {
+    private var binding:FragmentMediaLibraryBinding? = null
+    private lateinit var tabMediator: TabLayoutMediator
+    private val viewModel: MediaLibraryViewModel by viewModel<MediaLibraryViewModel>()
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentMediaLibraryBinding.inflate(inflater, container,false)
+        return binding?.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding?.run {
+
+            viewPager.adapter = ViewPagerAdapter(childFragmentManager, lifecycle)
+
+            tabMediator = TabLayoutMediator(tabLayout, viewPager) {tab, position ->
+                when(position){
+                    0 -> tab.text = getString(R.string.likeMusic)
+                    1 -> tab.text = getString(R.string.playListMusic)
+                }
+            }
+
+            tabMediator.attach()
+        }
+
+        viewModel.theme.observe(viewLifecycleOwner) {
+            AppCompatDelegate.setDefaultNightMode(it)
+        }
+
+        viewModel.requestTheme()
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        binding = null
+        tabMediator.detach()
+    }
+}
