+++ app/src/main/java/com/example/playlistmaker/media/presentation/view_model/MediaLibraryViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.main.presentation.view_model
+package com.example.playlistmaker.media.presentation.view_model
 
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.lifecycle.LiveData
@@ -6,9 +6,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.playlistmaker.settings.domain.SettingsInteractor
 
-
-class MainViewModel(private val settingsInteractor: SettingsInteractor) : ViewModel() {
-
+class MediaLibraryViewModel(private val settingsInteractor: SettingsInteractor) : ViewModel() {
     private val _theme = MutableLiveData(AppCompatDelegate.MODE_NIGHT_FOLLOW_SYSTEM)
     val theme: LiveData<Int>
         get() = _theme
@@ -17,5 +15,4 @@ class MainViewModel(private val settingsInteractor: SettingsInteractor) : ViewMo
     fun requestTheme() {
         _theme.postValue(settingsInteractor.getThemeSettings())
     }
-
 }
