+++ app/src/main/java/com/example/playlistmaker/ui/root/RootActivity.kt
@@ -1,0 +1,25 @@
+package com.example.playlistmaker.ui.root
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupWithNavController
+import com.example.playlistmaker.R
+import com.example.playlistmaker.databinding.ActivityRootBinding
+
+class RootActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityRootBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityRootBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.rootFragmentContainerView) as NavHostFragment
+        val navController = navHostFragment.navController
+
+        val bottomNavigationMenu =
+            binding.bottomNavigationView.setupWithNavController(navController)
+    }
+}
