+++ app/src/main/res/anim/slide_in_top.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <translate
+        android:fromYDelta="-100%p"
+        android:toYDelta="0"
+        android:duration="600" />
+</set>
