+++ app/src/main/res/anim/slide_out_top.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <translate
+        android:fromYDelta="0"
+        android:toYDelta="-100%p"
+        android:duration="600" />
+</set>
