+++ app/src/main/res/drawable/text_cursor.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <size android:width="@dimen/size_cursor"/>
+    <solid android:color="@color/blue_nav"/>
+</shape>
