+++ app/src/main/res/layout/fragment_media_library.xml
@@ -1,30 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/blackOrWhite"
-    android:orientation="vertical">
+    android:orientation="vertical"
+    tools:context=".media.presentation.fragment.MediaLibraryFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
 
-        <ImageView
-            android:id="@+id/toolBar"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_marginStart="@dimen/o_16dp"
-            android:layout_marginTop="@dimen/o_16dp"
-            android:layout_marginEnd="@dimen/o_16dp"
-            android:layout_marginBottom="@dimen/o_16dp"
-            android:src="@drawable/back_ic"
-            app:tint="@color/whiteOrBlack" />
-
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="match_parent"
-            android:layout_marginStart="@dimen/o_12dp"
+            android:layout_marginStart="@dimen/o_16dp"
             android:layout_marginTop="@dimen/o_12dp"
             android:layout_marginBottom="@dimen/o_12dp"
             android:fontFamily="@font/ys_display_medium"
