+++ app/src/main/res/layout/fragment_music_player.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/blackOrWhite"
     android:orientation="vertical"
-    tools:context=".player.presentation.activity.MusicPlayerActivity">
+    tools:context=".player.presentation.fragment.MusicPlayerFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/leftBorder"
