+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/searchFragment"
+        android:icon="@drawable/search_icon"
+        android:title="@string/searching" />
+
+    <item
+        android:id="@+id/mediaLibraryFragment"
+        android:icon="@drawable/media"
+        android:title="@string/media" />
+
+    <item
+        android:id="@+id/settingsFragment"
+        android:icon="@drawable/settings"
+        android:title="@string/settings" />
+
+</menu>
