+++ app/src/main/res/values/colors.xml
@@ -14,4 +14,5 @@
     <color name="greyOrBlack">#AEAFB4</color>
     <color name="grey">#AEAFB4</color>
     <color name="greyButton">#40AEAFB4</color>
+    <color name="blue_nav">#3772E7</color>
 </resources>
