+++ app/src/main/res/values/styles.xml
@@ -64,7 +64,7 @@
         <item name="android:fontFamily">@font/ys_display_medium</item>
         <item name="android:layout_marginTop">@dimen/o_12dp</item>
         <item name="android:layout_marginBottom">@dimen/o_12dp</item>
-        <item name="android:layout_marginStart">@dimen/o_12dp</item>
+        <item name="android:layout_marginStart">@dimen/o_16dp</item>
         <item name="android:textStyle">bold</item>
         <item name="android:textColor">@color/whiteOrBlack</item>
         <item name="android:textSize">@dimen/o_22sp</item>
